xAnomaly/serving/requirements.txt
numpy==1.21.5
paho_mqtt==1.6.1
requests==2.27.1
tensorflow==2.9.1
pandas==1.4.2
